pyproject.toml
[project]
name = "simpledocbuilder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docxcompose>=1.4.0",
    "docxtpl>=0.20.2",
    "imgkit>=1.2.3",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pypandoc>=1.16.2",
    "python-docx>=1.2.0",
]
